tests/wpt/tests/intersection-observer/v2/svg-reference-filter-occlusion.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width,initial-scale=1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #target {
    position: absolute;
    left: 10px;
    top: 10px;
    width: 100px;
    height: 100px;
    background: green;
  }
  svg {
    position: absolute;
    left: 110px;
    top: 110px;
    width: 100px;
    height: 100px;
    overflow: visible;
  }
  rect {
    pointer-events: bounding-box;
  }
</style>
<div id="target"></div>
<svg>
  <defs>
    <filter id="blur" x="-30%" y="-30%" width="160%" height="160%">
      <feGaussianBlur stdDeviation="30"/>
    </filter>
  </defs>
  <rect id="rect" x="0" y="0" width="100" height="100" fill="blue"/>
</svg>
<script>
async_test(t => {
  const target = document.getElementById("target");
  const rect = document.getElementById("rect");
  let observation_count = 0;

  const observer = new IntersectionObserver(t.step_func(entries => {
    entries.forEach(entry => {
      observation_count++;
      if (observation_count === 1) {
        assert_true(entry.isVisible, "Initially visible");
        // Trigger filter change to occlude target
        rect.style.filter = "url(#blur)";
      } else if (observation_count === 2) {
        assert_false(entry.isVisible, "Should be occluded by filter");
        observer.disconnect();
        t.done();
      }
    });
  }), {trackVisibility: true, delay: 100});

  observer.observe(target);
}, "IntersectionObserver observes visibility changes from SVG reference filter changes");
</script>
